Adds currency matcher to model specs
Shared code between have_customer_type_matcher and have_currency_matcher
is extracted to new enum_matcher class.
